separate home page into components
