+++ frontend/src/components/users/homePage.js
@@ -1,0 +1,13 @@
+import '../../App.css';
+import React from 'react';
+import QuestionHistory from "./questionHistory";
+
+
+
+function HomePage() {
+  return (
+    <QuestionHistory />
+  );
+}
+
+export default HomePage;
